requirements.txt
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
imdbpy==2021.4.18
psycopg2-binary==2.8.6
requests==2.25.1
Werkzeug==1.0.1